DATABASE MANAGEMENT SYSTEM ONLINE VEHICLES FOR SALE
A vehicle for sale system is a system that allows individuals or businesses to list vehicles for sale, and for potential buyers to browse and search for vehicles that meet their criteria. The system may be an online platform, such as a website or a mobile app, or it may be a physical location, such as a car dealership. Our system is based on website.